requirements.txt
asyncssh
bcrypt
colorama
pyyaml